refactor(util): share the symplectic inner product
anti_commute(MaskView, MaskView, MaskView, MaskView) had two copies:
optimizer/commutation.h's original and a local duplicate the planner
added because sampling/ had no other reason to depend on optimizer/.
Move it to util/symplectic.h instead, where both optimizer/ and
sampling/ can depend on it without depending on each other.
commutation.h includes the new header so its existing callers keep
compiling unchanged, and the planner now includes it too instead of
carrying its own copy.
